OVS-2663: Use default MaxRead/MaxWrite buffer size values
VMWare client seems to ask more than the MaxRead/MaxWrite values
we are already using.
This should not happen of course because the client has already taken
the values at mount time (FSINFO) and so is completely aware of the
server maximum read/write size.
Use the default values which are large enough and seem to solve
the problem.

Signed-off-by: Chrysostomos Nanakos <cnanakos@openvstorage.com>